Added onSetOrderReceived slot in WorkspacePage to remove received orders from IssuingOrdersModel (Employee application)

+++ Customer/client.h
@@ -42,7 +42,7 @@ signals:
 public slots:
     void onLogin(const QString& phone_number, const QString& password);
     void onRegister(const QString& phone_number, const QString& password, const QString& name);
-    void onMakeOrder(const QJsonArray& order_data, const size_t& total_cost);
+    void onMakeOrder(const QJsonArray& order_data, const int& total_cost);
     void onCancelOrder(const int &order_id, const QString& phone_number, const QString &receive_code);
     void onReadyRead();
     void onChangeName(const QString& new_name);
